Stage rules conformance corpus + capture runner (no captures)
Groundwork for the conformance-chain consolidation (#11), Firestore rules
phase, up to but NOT including any live network capture.

- New corpus: scripts/oracle/rules-corpus/firestore/. The 18 parity packs
  (12 stress + 6 round-1/2 fix-class) move out of the inline test files into
  typed Pack modules there. `types.ts` is the single home for the shared Pack
  type; harness.ts now re-exports it. Pack literals are byte-identical.
- The live-network parity tests (parity-stress, round-fix-classes) are
  unchanged in behavior — they now import STRESS_PACKS / FIX_CLASS_PACKS from
  the corpus instead of holding them inline.
- New capture runner: scripts/oracle/run-rules.ts. Reads the corpus and, when
  given PARITY_SA_BASE64 (same contract as the parity harness), posts each
  pack to the production Rules Test API via TestFirestoreRulesHandler, writing
  one observation per pack to scripts/oracle/observations/rules-firestore-*.json.
  Without the secret it is inert: prints the plan (packs, case counts, output
  paths, env var) and exits 0 with no network calls.
- Registered the `rules-firestore-` observation prefix in the compat registry.
- New replay skeleton: test/rules/oracle-conformance.test.ts. Data-driven over
  rules-firestore-* observations: replays each captured pack against the
  in-process simulator and asserts verdict-for-verdict. Skips (passing) while
  the observation set is empty; a captured observation with no corpus pack
  fails loudly, so completeness enforcement is live before any data exists.

No captures were run and no observation files were fabricated. The
emulator-diff harness and old parity tests are left in place for a later phase.
